requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django_extensions==3.2.1
easy_thumbnails==2.8.3
persiantools==3.0.1